requirements.txt
Braindecode==0.7
joblib==1.2.0
matplotlib==3.7.1
mne==1.3.0
numpy==1.23.5
pandas==1.5.3
scikit_learn==1.2.1
scipy==1.10.1
seaborn==0.13.0
skorch==0.12.1
torch==2.0.1
umap_learn==0.5.3
wfdb==4.1.2
